# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Local server"
        },
        {
            "url": "https://foodie-app.azurewebsites.net",
            "description": "Production server"
        }
    ],
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "foodie"
            }
        },
...
    },
    "security": [
        {
            "cookieAuth": []
        }
    ],
    "paths": {
...
        "/api/v1/{username}/posts": {
            "get": {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Get all posts by a user",
                "description": "Get all posts by a user",
                "operationId": "getPosts",
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "Username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "description": {
                                                "type": "string",
                                                "description": "Content",
                                                "example": "test"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/search": {
            "get": {
                "security": [
                    {
                        "cookieAuth": [
                            "foodie"
                        ]
                    }
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search for posts",
                "description": "Search for posts",
                "operationId": "searchPosts",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "description": {
                                                "type": "string",
                                                "description": "Content",
                                                "example": "test"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}